Modified hw project for Vivado 2021.1
No IP upgrades were required. The board part number was changed from 1.0
to 1.1 in order to be recognized by Vivado 2021.1.

In the top module the pins vn_in and vp_in were connected to GND in
order to solve the critical warning from the previous versions of
Vivado.

The xadc_wiz_0.xci was also modified automatically after the synthesis |
implementation in Vivado 2021.1.